+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## Version 4.1.1 (Under development)
 
+### App Center
+
+* **[Improvement]** Use ASWebAuthenticationSession for authentication on iOS 12 or later.
+
 ### App Center Distribute
 
 * **[Fix]** Fix `kMSACUpdateTokenRequestIdKey` never gets removed.
+
+___
 
 ## Version 4.1.0